+++ core/src/tracking.cpp
@@ -56,7 +56,7 @@ std::pair<std::vector<cv::Point2f>, std::vector<cv::Point2f> > tracking::LK(Fram
 
 	std::vector<cv::Point2f> vPs;
 	std::vector<cv::Point2f> vQs;
-	for (int i = 0; i < kpts1.size(); i++)
+	for (size_t i = 0; i < kpts1.size(); i++)
 	{
 		cv::Point2f pt1 = kpts1[i];
 		cv::Point2f pt2 = kpts2[i];
